Fix dependency audit and UI test ownership
