+++ package.json
@@ -23,6 +23,7 @@
 		"tevm": "1.0.0-rc.151"
 	},
 	"overrides": {
+		"fast-uri": "3.1.4",
 		"tmp": "0.2.7",
 		"ws": "8.21.0"
 	},
