+++ ui/package.json
@@ -25,6 +25,7 @@
 	},
 	"overrides": {
 		"brace-expansion": "5.0.7",
+		"fast-uri": "3.1.4",
 		"tmp": "0.2.7",
 		"ws": "8.21.0"
 	},
